requirements.txt
mss>=9.0.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
websockets>=12.0
PyYAML>=6.0
pytesseract>=0.3.10
